# docs/gpt-action-openapi.yaml
openapi: 3.1.0
info:
  title: timmeny-admin-os API
  version: 0.7.0
  description: GPT-safe API for the daily email review and for Timmeny Admin OS todos.
servers:
  - url: https://timmeny-admin-os-production.up.railway.app
paths:
...
  /review/runs/{run_id}/actions/prepare:
    post:
      operationId: prepareReviewActions
      summary: Turn approvals into exact, inspectable plans
      description: >-
        Preparing writes nothing to Gmail. It resolves each approved decision
        into the exact parameters that would be sent and a stable idempotency
        key, so the plan can be read before anything happens. Preparing the same
        approval twice returns the same action.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                capability_key:
                  type: string
                  description: Limit to one capability. Omit for the whole run.
                item_ids:
                  type: array
                  items:
                    type: string
                action_ids:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Actions prepared.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionsResponse"
        "409":
          description: >-
            The capability is not permitted to execute the action, or the
            parameters are not usable.
  /review/runs/{run_id}/actions:
    get:
      operationId: listReviewActions
      summary: Read every action on a run, with verification and failures
      security:
        - ApiKeyAuth: []
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
        - name: capability_key
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          description: Filter by lifecycle state. Use failed to find what needs a retry.
          schema:
            type: string
            enum: [approved, prepared, executed, verified, completed, failed]
      responses:
        "200":
          description: Actions on the run.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionsResponse"
  /review/runs/{run_id}/actions/{action_id}:
    get:
      operationId: readReviewAction
      summary: Read one action and its whole audit history
      security:
        - ApiKeyAuth: []
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
        - name: action_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The action, its verification, and every event recorded against it.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResponse"
        "404":
          description: No such action on this run.
  /review/runs/{run_id}/actions/execute:
    post:
      operationId: executeReviewActions
      summary: Execute prepared actions and verify each one
      description: >-
        The only route that changes the mailbox. Four gates stand in front of
        it: the capability must be allowed the action, separately permitted to
        execute it, GMAIL_WRITE_ENABLED must be on, and confirm must be true.
        Each action is read back from Gmail afterwards; an action Gmail does not
        confirm is left failed rather than reported as done.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - confirm
              properties:
                confirm:
                  type: boolean
                  description: Must be true. Executing writes to Gmail.
                capability_key:
                  type: string
                action_ids:
                  type: array
                  items:
                    type: string
                  description: Omit to execute everything prepared on the run.
      responses:
        "200":
          description: Actions executed and verified.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionsResponse"
        "400":
          description: confirm was not true.
        "409":
          description: Gmail writes are disabled, or the capability may not execute this.
        "502":
          description: Gmail could not be reached. The action is recorded as failed and can be retried.
        "503":
          description: Gmail is not configured.
  /review/runs/{run_id}/actions/{action_id}/retry:
    post:
      operationId: retryReviewAction
      summary: Retry one failed action without repeating its effect
      description: >-
        The effect is verified before it is attempted again, so a write that in
        fact landed before the connection dropped is adopted rather than made
        twice.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
        - name: action_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The action after the retry.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResponse"
        "404":
          description: No such action on this run.
  /review/runs/{run_id}/actions/{action_id}/verify:
    post:
      operationId: verifyReviewAction
      summary: Read an executed action back from Gmail again
      description: Writes nothing. Confirms that what was done is still what Gmail shows.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
        - name: action_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The action with fresh verification.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResponse"
  /review/runs/{run_id}/items/{item_id}/send-draft:
    post:
      operationId: approveSendDraft
      summary: Approve sending one exact draft
      description: >-
        Creating a draft never sends it. This approves the send of one specific
        draft, identified by both its draft id and the message id it contained
        when it was read back. Approving does not send either: the resulting
        action still has to be executed, and it refuses if the draft has changed
        since it was approved.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
        - name: item_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - draft_id
                - draft_message_id
                - confirm
              properties:
                draft_id:
                  type: string
                draft_message_id:
                  type: string
                  description: From the draft action's verification. Proves which draft was read.
                confirm:
                  type: boolean
      responses:
        "200":
          description: The send is approved, and not yet executed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResponse"
        "400":
          description: confirm was not true.
        "409":
          description: >-
            The capability may not send, or these ids are not the draft that was
            reviewed.
...
components:
  schemas:
...
    ActionEventResponse:
      type: object
      additionalProperties: true
      properties:
        sequence:
          type: integer
        event:
          type: string
          description: >-
            approved, prepared, execution_started, executed, adopted_existing,
            already_applied, verified, verification_failed, or failed.
        state_after:
          type: string
        detail:
          type: object
          additionalProperties: true
        external_ref:
          type: string
        at:
          type: string
          format: date-time
    ActionResponse:
      type: object
      additionalProperties: true
      properties:
        action_id:
          type: string
        item_id:
          type: string
        capability_key:
          type: string
        action:
          type: string
        state:
          type: string
          enum: [approved, prepared, executed, verified, completed, failed]
          description: >-
            approved is recorded intent and nothing more. prepared has exact
            parameters and has still written nothing. executed has reached
            Gmail. verified and completed mean Gmail was read back and agrees.
            failed is durable and retryable.
        params:
          type: object
          additionalProperties: true
        prepared_params:
          type: object
          additionalProperties: true
          description: Exactly what would be sent, resolved at preparation.
        idempotency_key:
          type: string
          description: Derived from the item, the action, and its parameters, so a retry cannot duplicate a write.
        thread_id:
          type: string
        external_kind:
          type: string
        external_ref:
          type: string
          description: The Gmail thread or draft the action touched.
        approval_kind:
          type: string
          enum: [human, automatable_rule]
        approved_by:
          type: string
        rule_id:
          type: string
          description: Set when a promoted rule approved this rather than a person.
        attempts:
          type: integer
        last_error:
          type: string
        verification:
          type: object
          additionalProperties: true
          description: What Gmail said when the action was read back.
        prepared_at:
          type: string
          format: date-time
        executed_at:
          type: string
          format: date-time
        verified_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        failed_at:
          type: string
          format: date-time
        events:
          type: array
          items:
            $ref: "#/components/schemas/ActionEventResponse"
    ActionsResponse:
      type: object
      additionalProperties: true
      properties:
        gmail_write_enabled:
          type: boolean
          description: The global kill switch. False means no action can reach the mailbox.
        counts:
          type: object
          additionalProperties: true
        actions:
          type: array
          items:
            $ref: "#/components/schemas/ActionResponse"
...
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer